Replace the hardcoded pixel width of Graph preview, with a static variable.
